refactor: 예외 패키지 전역 범위로 재지정
