--- src/main/java/roomescape/entity/exception/NullFieldExistException.java
@@ -1,9 +1,0 @@
-package roomescape.entity.exception;
-
-import roomescape.controller.exception.IllegalRequestException;
-
-public class NullFieldExistException extends IllegalRequestException {
-    public NullFieldExistException(String s) {
-        super(s);
-    }
-}
