--- src/main/java/roomescape/entity/exception/ReservationDateIsPastException.java
@@ -1,9 +1,0 @@
-package roomescape.entity.exception;
-
-import roomescape.controller.exception.IllegalRequestException;
-
-public class ReservationDateIsPastException extends IllegalRequestException {
-    public ReservationDateIsPastException(String s) {
-        super(s);
-    }
-}
