+++ src/main/java/roomescape/exception/GlobalExceptionHandler.java
@@ -1,12 +1,12 @@
-package roomescape.controller;
+package roomescape.exception;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import roomescape.controller.exception.IllegalRequestException;
 
 @RestControllerAdvice
-public class ApiExceptionAdvice {
+public class GlobalExceptionHandler {
+
     @ExceptionHandler(IllegalRequestException.class)
     public ResponseEntity<IllegalRequestException> badRequest(IllegalRequestException e) {
         return ResponseEntity.badRequest().body(e);
