+++ src/main/java/roomescape/exception/IllegalRequestException.java
@@ -1,4 +1,4 @@
-package roomescape.controller.exception;
+package roomescape.exception;
 
 public class IllegalRequestException extends RuntimeException {
     public IllegalRequestException(String message) {
