+++ src/main/java/roomescape/exception/IllegalRequestFormException.java
@@ -1,0 +1,7 @@
+package roomescape.exception;
+
+public class IllegalRequestFormException extends IllegalRequestException {
+    public IllegalRequestFormException(String s) {
+        super(s);
+    }
+}
