+++ src/main/java/roomescape/exception/IllegalReservationDateException.java
@@ -1,0 +1,8 @@
+package roomescape.exception;
+
+
+public class IllegalReservationDateException extends IllegalRequestException {
+    public IllegalReservationDateException(String s) {
+        super(s);
+    }
+}
