+++ src/main/java/roomescape/exception/IllegalReservationTimeException.java
@@ -1,0 +1,7 @@
+package roomescape.exception;
+
+public class IllegalReservationTimeException extends IllegalRequestException {
+    public IllegalReservationTimeException(String s) {
+        super(s);
+    }
+}
