--- src/main/java/roomescape/service/exception/SavedDataConflictWithRequestException.java
@@ -1,9 +1,0 @@
-package roomescape.service.exception;
-
-import roomescape.controller.exception.IllegalRequestException;
-
-public class SavedDataConflictWithRequestException extends IllegalRequestException {
-    public SavedDataConflictWithRequestException(String s) {
-        super(s);
-    }
-}
