+++ src/test/java/roomescape/entity/GameTimeTest.java
@@ -4,13 +4,13 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.entity.exception.NullFieldExistException;
+import roomescape.exception.IllegalRequestFormException;
 
 class GameTimeTest {
     @DisplayName("예약 가능 시각이 null인 경우 예외가 발생한다")
     @Test
     void createWithNullTest() {
         assertThatThrownBy(() -> new GameTime(null))
-                .isInstanceOf(NullFieldExistException.class);
+                .isInstanceOf(IllegalRequestFormException.class);
     }
 }
