+++ src/test/java/roomescape/entity/ReservationDateTest.java
@@ -5,15 +5,15 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.time.LocalDate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.entity.exception.NullFieldExistException;
-import roomescape.entity.exception.ReservationDateIsPastException;
+import roomescape.exception.IllegalRequestFormException;
+import roomescape.exception.IllegalReservationDateException;
 
 class ReservationDateTest {
     @DisplayName("날짜가 null인 경우 ReservationDate 생성에 실패한다")
     @Test
     void createReservationDateWithNullTest() {
         assertThatThrownBy(() -> new ReservationDate(null))
-                .isInstanceOf(NullFieldExistException.class);
+                .isInstanceOf(IllegalRequestFormException.class);
     }
 
     @DisplayName("날짜가 현재 날짜보다 이전 날짜인 경우 생성 시 예외가 발생한다")
@@ -22,6 +22,6 @@ class ReservationDateTest {
         LocalDate now = LocalDate.now();
         LocalDate passedDate = now.minusDays(1);
         assertThatThrownBy(() -> new ReservationDate(passedDate))
-                .isInstanceOf(ReservationDateIsPastException.class);
+                .isInstanceOf(IllegalReservationDateException.class);
     }
 }
